refactor: consolidate configuration code
