+++ virtool/caches/utils.py
@@ -4,7 +4,7 @@ Utilities used for working with cache files within analysis workflows.
 """
 from pathlib import Path
 
-from virtool.configuration.config import Config
+from virtool.config.cls import Config
 
 
 def join_cache_path(config: Config, cache_id: str) -> Path:
