+++ virtool/github.py
@@ -6,7 +6,7 @@ import aiohttp
 
 import virtool.errors
 import virtool.utils
-from virtool.configuration.config import Config
+from virtool.config.cls import Config
 from virtool.http.proxy import ProxyRequest
 
 logger = logging.getLogger(__name__)
